rename Toast to useToast
