+++ src/hooks/index.ts
@@ -9,4 +9,4 @@ export * from './useNavigation';
 export * from './useMediaQuery';
 export * from './usePrevious';
 export * from './useRecoilLoadCatch';
-export * from './Toast';
+export * from './useToast';
